mi4p: Multiple Imputation for Proteomics

A framework for multiple imputation for proteomics is proposed by Marie Chion, Christine Carapito and Frederic Bertrand (2021) doi:10.1371/journal.pcbi.1010420 . It is dedicated to dealing with multiple imputation for proteomics.

References

M. Chion, Ch. Carapito and F. Bertrand (2021). Accounting for multiple imputation-induced variability for differential analysis in mass spectrometry-based label-free quantitative proteomics. doi:10.1371/journal.pcbi.1010420

M. Chion, Ch. Carapito, F. Bertrand. Towards a more accurate differential analysis of multiple imputed proteomics data with mi4limma. Statistical Analysis of Proteomic Data: Methods and Tools, 2022. doi:10.1007/978-1-0716-1967-4_7

See also

Author

Maintainer: Frederic Bertrand frederic.bertrand@lecnam.net (ORCID)

Authors:

Marie Chion mariechion@protonmail.com (ORCID)

Christine Carapito ccarapito@unistra.fr (ORCID)

Other contributors:

Gordon Smyth [contributor]

Davis McCarthy [contributor]

Hélène Borges [contributor]

Thomas Burger [contributor]

Quentin Giai-Gianetto [contributor]

Samuel Wieczorek [contributor]