mm_peptides - peptide-level intensities for mouse

A dataset containing the protein and petide information and peptide-level intensities for 6 samples: 3 CG and 3 mCG groups. There are 69 proteins. The columns are as follows:

Usage

data(mm_peptides)

Format

A data frame with 1102 rows and 13 colummns, compiring 7 columns of metadata and 6 columns of peptide intensities. 69 proteins.

Details

Sequence - peptide sequence - randomly chosen from a larger list of sequences

MatchedID - numeric ID that links proteins in the two datasets, unnecessary if datasets are for the same species

ProtID - protein ID, artificial protein ID, eg. Prot1, Prot2, ...

GeneID - gene ID, artificial gene ID, eg. Gene1, Gene2, ...

ProtName - artificial Protein Name

ProtIDLong - long protein ID, full protein name, here artificially simulated

GeneIDLong - long gene ID, full gene name, here artificially simulated

CG1 - raw intensity column for sample 1 in CG group

CG2 - raw intensity column for sample 2 in CG group

CG3 - raw intensity column for sample 3 in CG group

mCG1 - raw intensity column for sample 1 in mCG group

mCG2 - raw intensity column for sample 2 in mCG group

mCG3 - raw intensity column for sample 3 in mCG group