Experimental design for the Exp1_R25_pept dataset

The data frame sTab contains the experimental design and gives few informations about the samples. They were obtained using the code pData(Exp1_R25_pept).

Format

A data frame with 6 observations on the following 3 variables.

Sample.name

a character vector

Condition

a character vector

Bio.Rep

a numeric vector

Source

The DAPARdata package.

Details

The DAPARdata's Exp1_R25_pept dataset is the final outcome of a quantitative mass spectrometry-based proteomic analysis of two samples containing different concentrations of 48 human proteins (UPS1 standard from Sigma-Aldrich) within a constant yeast background (see Giai Gianetto et al. (2016) for details). It contains the abundance values of the different human and yeast peptides identified and quantified in these two conditions. The two conditions represent the measured abundances of peptides when respectively 25 fmol and 10 fmol of UPS1 human proteins were mixed with the yeast extract before mass spectrometry analyses. This results in a concentration ratio of 2.5. Three technical replicates were acquired for each condition.

References

Cox J., Hein M.Y., Luber C.A., Paron I., Nagaraj N., Mann M. Accurate proteome-wide label-free quantification by delayed normalization and maximal peptide ratio extraction, termed MaxLFQ. Mol Cell Proteomics. 2014 Sep, 13(9):2513-26.

Giai Gianetto, Q., Combes, F., Ramus, C., Bruley, C., Coute, Y., Burger, T. (2016). Calibration plot for proteomics: A graphical tool to visually check the assumptions underlying FDR control in quantitative experiments. Proteomics, 16(1), 29-32.

Examples

data(sTab)
str(sTab)
#> 'data.frame':	6 obs. of  3 variables:
#>  $ Sample.name: chr  "Intensity_C_R1" "Intensity_C_R2" "Intensity_C_R3" "Intensity_D_R1" ...
#>  $ Condition  : chr  "25fmol" "25fmol" "25fmol" "10fmol" ...
#>  $ Bio.Rep    : int  1 2 3 4 5 6