MI-aware Modifed eBayes Function

Modified eBayes function to be used instead of the one, .ebayes, implemented in the limma package

hid.ebayes(
  fit,
  VarRubin,
  mod = TRUE,
  proportion = 0.01,
  stdev.coef.lim = c(0.1, 4),
  trend = FALSE,
  robust = FALSE,
  winsor.tail.p = c(0.05, 0.1)
)

Arguments

fit
an MArrayLM fitted model object produced by lmFit or contrasts.fit. For ebayes only, fit can alternatively be an unclassed list produced by lm.series, gls.series or mrlm containing components coefficients, stdev.unscaled, sigma and df.residual.

VarRubin
a variance-covariance matrix.

mod
TRUE (not used at the moment)

proportion
numeric value between 0 and 1, assumed proportion of genes which are differentially expressed

stdev.coef.lim
numeric vector of length 2, assumed lower and upper limits for the standard deviation of log2-fold-changes for differentially expressed genes

trend
logical, should an intensity-trend be allowed for the prior variance? Default is that the prior variance is constant.

robust
logical, should the estimation of df.prior and var.prior be robustified against outlier sample variances?

winsor.tail.p
numeric vector of length 1 or 2, giving left and right tail proportions of x to Winsorize. Used only when robust=TRUE.

Value

eBayes produces an object of class MArrayLM (see MArrayLM-class) containing everything found in fit plus the following added components:

t

numeric matrix of moderated t-statistics.

p.value

numeric matrix of two-sided p-values corresponding to the t-statistics.

lods

numeric matrix giving the log-odds of differential expression (on the natural log scale).

s2.prior

estimated prior value for sigma^2. A row-wise vector if covariate is non-NULL, otherwise a single value.

df.prior

degrees of freedom associated with s2.prior. A row-wise vector if robust=TRUE, otherwise a single value.

df.total

row-wise numeric vector giving the total degrees of freedom associated with the t-statistics for each gene. Equal to df.prior+df.residual or sum(df.residual), whichever is smaller.

s2.post

row-wise numeric vector giving the posterior values for sigma^2.

var.prior

column-wise numeric vector giving estimated prior values for the variance of the log2-fold-changes for differentially expressed gene for each constrast. Used for evaluating lods.

F

row-wise numeric vector of moderated F-statistics for testing all contrasts defined by the columns of fit simultaneously equal to zero.

F.p.value

row-wise numeric vector giving p-values corresponding to F.

The matrices t, p.value and lods have the same dimensions as the input object fit, with rows corresponding to genes and columns to coefficients or contrasts. The vectors s2.prior, df.prior, df.total, F and F.p.value correspond to rows, with length equal to the number of genes. The vector var.prior corresponds to columns, with length equal to the number of contrasts. If s2.prior or df.prior have length 1, then the same value applies to all genes.

s2.prior, df.prior and var.prior contain empirical Bayes hyperparameters used to obtain df.total, s2.post and lods.

Author

Modified by M. Chion and F. Bertrand. Original by Gordon Smyth and Davis McCarthy

Examples

library(mi4p)
data(datasim)
datasim_imp <- multi.impute(data = datasim[,-1], conditions = 
attr(datasim,"metadata")$Condition, method = "MLE")
VarRubin.matrix <- rubin2.all(datasim_imp[1:5,,],
attr(datasim,"metadata")$Condition)
set.seed(2016)
sigma2 <- 0.05 / rchisq(100, df=10) * 10
y <- datasim_imp[,,1]
design <- cbind(Intercept=1,Group=as.numeric(
attr(datasim,"metadata")$Condition)-1)
fit.model <- limma::lmFit(y,design)
hid.ebayes(fit=fit.model,VarRubin.matrix[[1]])

#> Warning: Zero sample variances detected, have been offset away from zero

#> $df.prior
#> [1] 0.2886034
#> 
#> $s2.prior
#> [1] 4.391122e-07
#> 
#> $s2.post
#>   [1] 3.526045e-02 1.528958e-08 1.528958e-08 3.526045e-02           NA
#>   [6]           NA           NA           NA           NA           NA
#>  [11]           NA           NA           NA           NA           NA
#>  [16]           NA           NA           NA           NA           NA
#>  [21]           NA           NA           NA           NA           NA
#>  [26]           NA           NA           NA           NA           NA
#>  [31]           NA           NA           NA           NA           NA
#>  [36]           NA           NA           NA           NA           NA
#>  [41]           NA           NA           NA           NA           NA
#>  [46]           NA           NA           NA           NA           NA
#>  [51]           NA           NA           NA           NA           NA
#>  [56]           NA           NA           NA           NA           NA
#>  [61]           NA           NA           NA           NA           NA
#>  [66]           NA           NA           NA           NA           NA
#>  [71]           NA           NA           NA           NA           NA
#>  [76]           NA           NA           NA           NA           NA
#>  [81]           NA           NA           NA           NA           NA
#>  [86]           NA           NA           NA           NA           NA
#>  [91]           NA           NA           NA           NA           NA
#>  [96]           NA           NA           NA           NA           NA
#> [101]           NA           NA           NA           NA           NA
#> [106]           NA           NA           NA           NA           NA
#> [111]           NA           NA           NA           NA           NA
#> [116]           NA           NA           NA           NA           NA
#> [121]           NA           NA           NA           NA           NA
#> [126]           NA           NA           NA           NA           NA
#> [131]           NA           NA           NA           NA           NA
#> [136]           NA           NA           NA           NA           NA
#> [141]           NA           NA           NA           NA           NA
#> [146]           NA           NA           NA           NA           NA
#> [151]           NA           NA           NA           NA           NA
#> [156]           NA           NA           NA           NA           NA
#> [161]           NA           NA           NA           NA           NA
#> [166]           NA           NA           NA           NA           NA
#> [171]           NA           NA           NA           NA           NA
#> [176]           NA           NA           NA           NA           NA
#> [181]           NA           NA           NA           NA           NA
#> [186]           NA           NA           NA           NA           NA
#> [191]           NA           NA           NA           NA           NA
#> [196]           NA           NA           NA           NA           NA
#> 
#> $t
#>          Intercept        Group
#>   [1,]    1179.208     852.2383
#>   [2,] 1811647.135 1269400.9619
#>   [3,] 1801623.098 1284808.0990
#>   [4,]    1191.913     842.6404
#>   [5,]          NA           NA
#>   [6,]          NA           NA
#>   [7,]          NA           NA
#>   [8,]          NA           NA
#>   [9,]          NA           NA
#>  [10,]          NA           NA
#>  [11,]          NA           NA
#>  [12,]          NA           NA
#>  [13,]          NA           NA
#>  [14,]          NA           NA
#>  [15,]          NA           NA
#>  [16,]          NA           NA
#>  [17,]          NA           NA
#>  [18,]          NA           NA
#>  [19,]          NA           NA
#>  [20,]          NA           NA
#>  [21,]          NA           NA
#>  [22,]          NA           NA
#>  [23,]          NA           NA
#>  [24,]          NA           NA
#>  [25,]          NA           NA
#>  [26,]          NA           NA
#>  [27,]          NA           NA
#>  [28,]          NA           NA
#>  [29,]          NA           NA
#>  [30,]          NA           NA
#>  [31,]          NA           NA
#>  [32,]          NA           NA
#>  [33,]          NA           NA
#>  [34,]          NA           NA
#>  [35,]          NA           NA
#>  [36,]          NA           NA
#>  [37,]          NA           NA
#>  [38,]          NA           NA
#>  [39,]          NA           NA
#>  [40,]          NA           NA
#>  [41,]          NA           NA
#>  [42,]          NA           NA
#>  [43,]          NA           NA
#>  [44,]          NA           NA
#>  [45,]          NA           NA
#>  [46,]          NA           NA
#>  [47,]          NA           NA
#>  [48,]          NA           NA
#>  [49,]          NA           NA
#>  [50,]          NA           NA
#>  [51,]          NA           NA
#>  [52,]          NA           NA
#>  [53,]          NA           NA
#>  [54,]          NA           NA
#>  [55,]          NA           NA
#>  [56,]          NA           NA
#>  [57,]          NA           NA
#>  [58,]          NA           NA
#>  [59,]          NA           NA
#>  [60,]          NA           NA
#>  [61,]          NA           NA
#>  [62,]          NA           NA
#>  [63,]          NA           NA
#>  [64,]          NA           NA
#>  [65,]          NA           NA
#>  [66,]          NA           NA
#>  [67,]          NA           NA
#>  [68,]          NA           NA
#>  [69,]          NA           NA
#>  [70,]          NA           NA
#>  [71,]          NA           NA
#>  [72,]          NA           NA
#>  [73,]          NA           NA
#>  [74,]          NA           NA
#>  [75,]          NA           NA
#>  [76,]          NA           NA
#>  [77,]          NA           NA
#>  [78,]          NA           NA
#>  [79,]          NA           NA
#>  [80,]          NA           NA
#>  [81,]          NA           NA
#>  [82,]          NA           NA
#>  [83,]          NA           NA
#>  [84,]          NA           NA
#>  [85,]          NA           NA
#>  [86,]          NA           NA
#>  [87,]          NA           NA
#>  [88,]          NA           NA
#>  [89,]          NA           NA
#>  [90,]          NA           NA
#>  [91,]          NA           NA
#>  [92,]          NA           NA
#>  [93,]          NA           NA
#>  [94,]          NA           NA
#>  [95,]          NA           NA
#>  [96,]          NA           NA
#>  [97,]          NA           NA
#>  [98,]          NA           NA
#>  [99,]          NA           NA
#> [100,]          NA           NA
#> [101,]          NA           NA
#> [102,]          NA           NA
#> [103,]          NA           NA
#> [104,]          NA           NA
#> [105,]          NA           NA
#> [106,]          NA           NA
#> [107,]          NA           NA
#> [108,]          NA           NA
#> [109,]          NA           NA
#> [110,]          NA           NA
#> [111,]          NA           NA
#> [112,]          NA           NA
#> [113,]          NA           NA
#> [114,]          NA           NA
#> [115,]          NA           NA
#> [116,]          NA           NA
#> [117,]          NA           NA
#> [118,]          NA           NA
#> [119,]          NA           NA
#> [120,]          NA           NA
#> [121,]          NA           NA
#> [122,]          NA           NA
#> [123,]          NA           NA
#> [124,]          NA           NA
#> [125,]          NA           NA
#> [126,]          NA           NA
#> [127,]          NA           NA
#> [128,]          NA           NA
#> [129,]          NA           NA
#> [130,]          NA           NA
#> [131,]          NA           NA
#> [132,]          NA           NA
#> [133,]          NA           NA
#> [134,]          NA           NA
#> [135,]          NA           NA
#> [136,]          NA           NA
#> [137,]          NA           NA
#> [138,]          NA           NA
#> [139,]          NA           NA
#> [140,]          NA           NA
#> [141,]          NA           NA
#> [142,]          NA           NA
#> [143,]          NA           NA
#> [144,]          NA           NA
#> [145,]          NA           NA
#> [146,]          NA           NA
#> [147,]          NA           NA
#> [148,]          NA           NA
#> [149,]          NA           NA
#> [150,]          NA           NA
#> [151,]          NA           NA
#> [152,]          NA           NA
#> [153,]          NA           NA
#> [154,]          NA           NA
#> [155,]          NA           NA
#> [156,]          NA           NA
#> [157,]          NA           NA
#> [158,]          NA           NA
#> [159,]          NA           NA
#> [160,]          NA           NA
#> [161,]          NA           NA
#> [162,]          NA           NA
#> [163,]          NA           NA
#> [164,]          NA           NA
#> [165,]          NA           NA
#> [166,]          NA           NA
#> [167,]          NA           NA
#> [168,]          NA           NA
#> [169,]          NA           NA
#> [170,]          NA           NA
#> [171,]          NA           NA
#> [172,]          NA           NA
#> [173,]          NA           NA
#> [174,]          NA           NA
#> [175,]          NA           NA
#> [176,]          NA           NA
#> [177,]          NA           NA
#> [178,]          NA           NA
#> [179,]          NA           NA
#> [180,]          NA           NA
#> [181,]          NA           NA
#> [182,]          NA           NA
#> [183,]          NA           NA
#> [184,]          NA           NA
#> [185,]          NA           NA
#> [186,]          NA           NA
#> [187,]          NA           NA
#> [188,]          NA           NA
#> [189,]          NA           NA
#> [190,]          NA           NA
#> [191,]          NA           NA
#> [192,]          NA           NA
#> [193,]          NA           NA
#> [194,]          NA           NA
#> [195,]          NA           NA
#> [196,]          NA           NA
#> [197,]          NA           NA
#> [198,]          NA           NA
#> [199,]          NA           NA
#> [200,]          NA           NA
#> 
#> $df.total
#>   [1] 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603
#>   [9] 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603
#>  [17] 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603
#>  [25] 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603
#>  [33] 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603
#>  [41] 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603
#>  [49] 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603
#>  [57] 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603
#>  [65] 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603
#>  [73] 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603
#>  [81] 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603
#>  [89] 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603
#>  [97] 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603
#> [105] 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603
#> [113] 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603
#> [121] 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603
#> [129] 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603
#> [137] 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603
#> [145] 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603
#> [153] 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603
#> [161] 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603
#> [169] 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603
#> [177] 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603
#> [185] 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603
#> [193] 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603 8.288603
#> 
#> $p.value
#>           Intercept        Group
#>   [1,] 5.982632e-23 8.827114e-22
#>   [2,] 2.319558e-49 4.423698e-48
#>   [3,] 2.428723e-49 4.002744e-48
#>   [4,] 5.474126e-23 9.695892e-22
#>   [5,]           NA           NA
#>   [6,]           NA           NA
#>   [7,]           NA           NA
#>   [8,]           NA           NA
#>   [9,]           NA           NA
#>  [10,]           NA           NA
#>  [11,]           NA           NA
#>  [12,]           NA           NA
#>  [13,]           NA           NA
#>  [14,]           NA           NA
#>  [15,]           NA           NA
#>  [16,]           NA           NA
#>  [17,]           NA           NA
#>  [18,]           NA           NA
#>  [19,]           NA           NA
#>  [20,]           NA           NA
#>  [21,]           NA           NA
#>  [22,]           NA           NA
#>  [23,]           NA           NA
#>  [24,]           NA           NA
#>  [25,]           NA           NA
#>  [26,]           NA           NA
#>  [27,]           NA           NA
#>  [28,]           NA           NA
#>  [29,]           NA           NA
#>  [30,]           NA           NA
#>  [31,]           NA           NA
#>  [32,]           NA           NA
#>  [33,]           NA           NA
#>  [34,]           NA           NA
#>  [35,]           NA           NA
#>  [36,]           NA           NA
#>  [37,]           NA           NA
#>  [38,]           NA           NA
#>  [39,]           NA           NA
#>  [40,]           NA           NA
#>  [41,]           NA           NA
#>  [42,]           NA           NA
#>  [43,]           NA           NA
#>  [44,]           NA           NA
#>  [45,]           NA           NA
#>  [46,]           NA           NA
#>  [47,]           NA           NA
#>  [48,]           NA           NA
#>  [49,]           NA           NA
#>  [50,]           NA           NA
#>  [51,]           NA           NA
#>  [52,]           NA           NA
#>  [53,]           NA           NA
#>  [54,]           NA           NA
#>  [55,]           NA           NA
#>  [56,]           NA           NA
#>  [57,]           NA           NA
#>  [58,]           NA           NA
#>  [59,]           NA           NA
#>  [60,]           NA           NA
#>  [61,]           NA           NA
#>  [62,]           NA           NA
#>  [63,]           NA           NA
#>  [64,]           NA           NA
#>  [65,]           NA           NA
#>  [66,]           NA           NA
#>  [67,]           NA           NA
#>  [68,]           NA           NA
#>  [69,]           NA           NA
#>  [70,]           NA           NA
#>  [71,]           NA           NA
#>  [72,]           NA           NA
#>  [73,]           NA           NA
#>  [74,]           NA           NA
#>  [75,]           NA           NA
#>  [76,]           NA           NA
#>  [77,]           NA           NA
#>  [78,]           NA           NA
#>  [79,]           NA           NA
#>  [80,]           NA           NA
#>  [81,]           NA           NA
#>  [82,]           NA           NA
#>  [83,]           NA           NA
#>  [84,]           NA           NA
#>  [85,]           NA           NA
#>  [86,]           NA           NA
#>  [87,]           NA           NA
#>  [88,]           NA           NA
#>  [89,]           NA           NA
#>  [90,]           NA           NA
#>  [91,]           NA           NA
#>  [92,]           NA           NA
#>  [93,]           NA           NA
#>  [94,]           NA           NA
#>  [95,]           NA           NA
#>  [96,]           NA           NA
#>  [97,]           NA           NA
#>  [98,]           NA           NA
#>  [99,]           NA           NA
#> [100,]           NA           NA
#> [101,]           NA           NA
#> [102,]           NA           NA
#> [103,]           NA           NA
#> [104,]           NA           NA
#> [105,]           NA           NA
#> [106,]           NA           NA
#> [107,]           NA           NA
#> [108,]           NA           NA
#> [109,]           NA           NA
#> [110,]           NA           NA
#> [111,]           NA           NA
#> [112,]           NA           NA
#> [113,]           NA           NA
#> [114,]           NA           NA
#> [115,]           NA           NA
#> [116,]           NA           NA
#> [117,]           NA           NA
#> [118,]           NA           NA
#> [119,]           NA           NA
#> [120,]           NA           NA
#> [121,]           NA           NA
#> [122,]           NA           NA
#> [123,]           NA           NA
#> [124,]           NA           NA
#> [125,]           NA           NA
#> [126,]           NA           NA
#> [127,]           NA           NA
#> [128,]           NA           NA
#> [129,]           NA           NA
#> [130,]           NA           NA
#> [131,]           NA           NA
#> [132,]           NA           NA
#> [133,]           NA           NA
#> [134,]           NA           NA
#> [135,]           NA           NA
#> [136,]           NA           NA
#> [137,]           NA           NA
#> [138,]           NA           NA
#> [139,]           NA           NA
#> [140,]           NA           NA
#> [141,]           NA           NA
#> [142,]           NA           NA
#> [143,]           NA           NA
#> [144,]           NA           NA
#> [145,]           NA           NA
#> [146,]           NA           NA
#> [147,]           NA           NA
#> [148,]           NA           NA
#> [149,]           NA           NA
#> [150,]           NA           NA
#> [151,]           NA           NA
#> [152,]           NA           NA
#> [153,]           NA           NA
#> [154,]           NA           NA
#> [155,]           NA           NA
#> [156,]           NA           NA
#> [157,]           NA           NA
#> [158,]           NA           NA
#> [159,]           NA           NA
#> [160,]           NA           NA
#> [161,]           NA           NA
#> [162,]           NA           NA
#> [163,]           NA           NA
#> [164,]           NA           NA
#> [165,]           NA           NA
#> [166,]           NA           NA
#> [167,]           NA           NA
#> [168,]           NA           NA
#> [169,]           NA           NA
#> [170,]           NA           NA
#> [171,]           NA           NA
#> [172,]           NA           NA
#> [173,]           NA           NA
#> [174,]           NA           NA
#> [175,]           NA           NA
#> [176,]           NA           NA
#> [177,]           NA           NA
#> [178,]           NA           NA
#> [179,]           NA           NA
#> [180,]           NA           NA
#> [181,]           NA           NA
#> [182,]           NA           NA
#> [183,]           NA           NA
#> [184,]           NA           NA
#> [185,]           NA           NA
#> [186,]           NA           NA
#> [187,]           NA           NA
#> [188,]           NA           NA
#> [189,]           NA           NA
#> [190,]           NA           NA
#> [191,]           NA           NA
#> [192,]           NA           NA
#> [193,]           NA           NA
#> [194,]           NA           NA
#> [195,]           NA           NA
#> [196,]           NA           NA
#> [197,]           NA           NA
#> [198,]           NA           NA
#> [199,]           NA           NA
#> [200,]           NA           NA
#> 
#> $var.prior
#> [1] 36437154 36437154
#> 
#> $lods
#>        Intercept    Group
#>   [1,]  41.76277 39.09288
#>   [2,]  74.22959 71.35694
#>   [3,]  74.22957 71.35699
#>   [4,]  41.86223 38.98778
#>   [5,]        NA       NA
#>   [6,]        NA       NA
#>   [7,]        NA       NA
#>   [8,]        NA       NA
#>   [9,]        NA       NA
#>  [10,]        NA       NA
#>  [11,]        NA       NA
#>  [12,]        NA       NA
#>  [13,]        NA       NA
#>  [14,]        NA       NA
#>  [15,]        NA       NA
#>  [16,]        NA       NA
#>  [17,]        NA       NA
#>  [18,]        NA       NA
#>  [19,]        NA       NA
#>  [20,]        NA       NA
#>  [21,]        NA       NA
#>  [22,]        NA       NA
#>  [23,]        NA       NA
#>  [24,]        NA       NA
#>  [25,]        NA       NA
#>  [26,]        NA       NA
#>  [27,]        NA       NA
#>  [28,]        NA       NA
#>  [29,]        NA       NA
#>  [30,]        NA       NA
#>  [31,]        NA       NA
#>  [32,]        NA       NA
#>  [33,]        NA       NA
#>  [34,]        NA       NA
#>  [35,]        NA       NA
#>  [36,]        NA       NA
#>  [37,]        NA       NA
#>  [38,]        NA       NA
#>  [39,]        NA       NA
#>  [40,]        NA       NA
#>  [41,]        NA       NA
#>  [42,]        NA       NA
#>  [43,]        NA       NA
#>  [44,]        NA       NA
#>  [45,]        NA       NA
#>  [46,]        NA       NA
#>  [47,]        NA       NA
#>  [48,]        NA       NA
#>  [49,]        NA       NA
#>  [50,]        NA       NA
#>  [51,]        NA       NA
#>  [52,]        NA       NA
#>  [53,]        NA       NA
#>  [54,]        NA       NA
#>  [55,]        NA       NA
#>  [56,]        NA       NA
#>  [57,]        NA       NA
#>  [58,]        NA       NA
#>  [59,]        NA       NA
#>  [60,]        NA       NA
#>  [61,]        NA       NA
#>  [62,]        NA       NA
#>  [63,]        NA       NA
#>  [64,]        NA       NA
#>  [65,]        NA       NA
#>  [66,]        NA       NA
#>  [67,]        NA       NA
#>  [68,]        NA       NA
#>  [69,]        NA       NA
#>  [70,]        NA       NA
#>  [71,]        NA       NA
#>  [72,]        NA       NA
#>  [73,]        NA       NA
#>  [74,]        NA       NA
#>  [75,]        NA       NA
#>  [76,]        NA       NA
#>  [77,]        NA       NA
#>  [78,]        NA       NA
#>  [79,]        NA       NA
#>  [80,]        NA       NA
#>  [81,]        NA       NA
#>  [82,]        NA       NA
#>  [83,]        NA       NA
#>  [84,]        NA       NA
#>  [85,]        NA       NA
#>  [86,]        NA       NA
#>  [87,]        NA       NA
#>  [88,]        NA       NA
#>  [89,]        NA       NA
#>  [90,]        NA       NA
#>  [91,]        NA       NA
#>  [92,]        NA       NA
#>  [93,]        NA       NA
#>  [94,]        NA       NA
#>  [95,]        NA       NA
#>  [96,]        NA       NA
#>  [97,]        NA       NA
#>  [98,]        NA       NA
#>  [99,]        NA       NA
#> [100,]        NA       NA
#> [101,]        NA       NA
#> [102,]        NA       NA
#> [103,]        NA       NA
#> [104,]        NA       NA
#> [105,]        NA       NA
#> [106,]        NA       NA
#> [107,]        NA       NA
#> [108,]        NA       NA
#> [109,]        NA       NA
#> [110,]        NA       NA
#> [111,]        NA       NA
#> [112,]        NA       NA
#> [113,]        NA       NA
#> [114,]        NA       NA
#> [115,]        NA       NA
#> [116,]        NA       NA
#> [117,]        NA       NA
#> [118,]        NA       NA
#> [119,]        NA       NA
#> [120,]        NA       NA
#> [121,]        NA       NA
#> [122,]        NA       NA
#> [123,]        NA       NA
#> [124,]        NA       NA
#> [125,]        NA       NA
#> [126,]        NA       NA
#> [127,]        NA       NA
#> [128,]        NA       NA
#> [129,]        NA       NA
#> [130,]        NA       NA
#> [131,]        NA       NA
#> [132,]        NA       NA
#> [133,]        NA       NA
#> [134,]        NA       NA
#> [135,]        NA       NA
#> [136,]        NA       NA
#> [137,]        NA       NA
#> [138,]        NA       NA
#> [139,]        NA       NA
#> [140,]        NA       NA
#> [141,]        NA       NA
#> [142,]        NA       NA
#> [143,]        NA       NA
#> [144,]        NA       NA
#> [145,]        NA       NA
#> [146,]        NA       NA
#> [147,]        NA       NA
#> [148,]        NA       NA
#> [149,]        NA       NA
#> [150,]        NA       NA
#> [151,]        NA       NA
#> [152,]        NA       NA
#> [153,]        NA       NA
#> [154,]        NA       NA
#> [155,]        NA       NA
#> [156,]        NA       NA
#> [157,]        NA       NA
#> [158,]        NA       NA
#> [159,]        NA       NA
#> [160,]        NA       NA
#> [161,]        NA       NA
#> [162,]        NA       NA
#> [163,]        NA       NA
#> [164,]        NA       NA
#> [165,]        NA       NA
#> [166,]        NA       NA
#> [167,]        NA       NA
#> [168,]        NA       NA
#> [169,]        NA       NA
#> [170,]        NA       NA
#> [171,]        NA       NA
#> [172,]        NA       NA
#> [173,]        NA       NA
#> [174,]        NA       NA
#> [175,]        NA       NA
#> [176,]        NA       NA
#> [177,]        NA       NA
#> [178,]        NA       NA
#> [179,]        NA       NA
#> [180,]        NA       NA
#> [181,]        NA       NA
#> [182,]        NA       NA
#> [183,]        NA       NA
#> [184,]        NA       NA
#> [185,]        NA       NA
#> [186,]        NA       NA
#> [187,]        NA       NA
#> [188,]        NA       NA
#> [189,]        NA       NA
#> [190,]        NA       NA
#> [191,]        NA       NA
#> [192,]        NA       NA
#> [193,]        NA       NA
#> [194,]        NA       NA
#> [195,]        NA       NA
#> [196,]        NA       NA
#> [197,]        NA       NA
#> [198,]        NA       NA
#> [199,]        NA       NA
#> [200,]        NA       NA
#>