Data simulation function

Function to simulate benchmark datasets.

protdatasim(
  iii = 1,
  nobs = 200,
  nobs1 = 10,
  ng1 = 5,
  ng2 = 5,
  mg1 = 100,
  mg2 = 200,
  dispg1 = 1,
  dispg2 = 1
)

Arguments

iii
A parameter useful to loop over for simulated lists of datasets. It has no effect.

nobs
Number of peptides

nobs1
Number of peptides with differential expressions between the two conditions

ng1
Number of biological replicates in condition A

ng2
Number of biological replicates in condition B

mg1
Mean in condition A

mg2
Mean in condition B

dispg1
Dispersion in condition A

dispg2
Dispersion in condition B

Value

A data frame with the simulated and attribute metadata.

References

M. Chion, Ch. Carapito and F. Bertrand (2021). Accounting for multiple imputation-induced variability for differential analysis in mass spectrometry-based label-free quantitative proteomics. arxiv:2108.07086. https://arxiv.org/abs/2108.07086.

Examples

data_sim <- protdatasim()
attr(data_sim,"metadata")

#>    Sample.name Condition Bio.Rep
#> 1           X1         A       1
#> 2           X2         A       2
#> 3           X3         A       3
#> 4           X4         A       4
#> 5           X5         A       5
#> 6           X6         B       6
#> 7           X7         B       7
#> 8           X8         B       8
#> 9           X9         B       9
#> 10         X10         B      10

norm.200.m100.sd1.vs.m200.sd1_list <- lapply(1:100, protdatasim)
attr(norm.200.m100.sd1.vs.m200.sd1_list[[1]],"metadata")

#>    Sample.name Condition Bio.Rep
#> 1           X1         A       1
#> 2           X2         A       2
#> 3           X3         A       3
#> 4           X4         A       4
#> 5           X5         A       5
#> 6           X6         B       6
#> 7           X7         B       7
#> 8           X8         B       8
#> 9           X9         B       9
#> 10         X10         B      10